Restructure the root directory
