--- main-app/requirememt.txt
@@ -1,15 +1,0 @@
-aniso8601==2.0.1
-attrs==17.4.0
-click==6.7
-colorama==0.3.9
-Flask==0.12.2
-Flask-RESTful==0.3.6
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-pluggy==0.6.0
-py==1.5.2
-pytest==3.4.1
-pytz==2018.3
-six==1.11.0
-Werkzeug==0.14.1
