+++ src/resources/business.py
@@ -2,7 +2,7 @@ from flask import jsonify, Blueprint
 
 from flask.ext.restful import (Resource, Api, reqparse)
 
-from models.business import Business
+from src.models.business import Business
 
 
 class BusinessRecord(Resource):
