+++ tests/test_business_model.py
@@ -1,6 +1,6 @@
 import unittest
 
-from models.business import Business
+from src.models.business import Business
 
 
 class CreateBusinessTest(unittest.TestCase):
