+++ tests/test_flask_config.py
@@ -1,6 +1,5 @@
 import unittest
 
-
 from app import app
